% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testresults.R
\docType{data}
\name{allTestResults}
\alias{allTestResults}
\title{Results from running testing/Testprogram.R}
\format{list of lists}
\source{
A run of the program Testprogram.R, which can be found in the 
"testing" directory.
}
\usage{
data(allTestResults)
}
\description{
This data set contains the results from running the program 
testing/Testprogram.R. See testing/Testprogram.R and especially 
R/testing_power_calc_vs_power_sim.R for better understanding of the results.
}
\keyword{datasets}

