% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testresults.R
\docType{data}
\name{overviewTestResults}
\alias{overviewTestResults}
\title{Overview of the results from running testing/Testprogram.R}
\format{list of lists}
\source{
A run of the program Testprogram.R, which can be found in the 
"testing" directory.
}
\usage{
data(overviewTestResults)
}
\description{
This data set contains some descriptive statistics 
giving an overview of the results from running the program 
testing/Testprogram.R. See testing/Testprogram.R and especially 
R/testing_power_calc_vs_power_sim.R for better understanding of how
the overview was constructed and what it describes.
}
\keyword{datasets}

