% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_t_test_clustered.R
\name{power.t.test.clustered}
\alias{power.t.test.clustered}
\title{Power of the adjusted t test for cluster randomized designs}
\usage{
power.t.test.clustered(clusters_group_1 = NULL,
  clusters_group_2 = clusters_group_1, delta = NULL, sd = NULL,
  rho = NULL, power = NULL, alternative = c("one.sided", "two.sided"),
  sig_level = 0.05, sample_size_calc = FALSE, clusters_pr_group = NULL,
  cluster_size = NULL)
}
\arguments{
\item{delta}{the difference between the group means (the treatment effect).}

\item{sd}{the overall standard deviation of the responses, which is assumed 
to be equal in both groups.}

\item{rho}{the intracluster correlation coeffiecient, which is assumed to 
be equal in both groups.}

\item{power}{the power of the cluster adjusted t test.}

\item{alternative}{has to be set to one of two strings, either 
\code{"one.sided"} or \code{"two.sided"}, to say whether the alternative 
hypothesis is one-sided or two-sided.}

\item{sig_level}{the significance level of the t test. If it is not 
specified, it has default value \code{0.05}.}

\item{sample_size_calc}{has to be set to \code{TRUE} or \code{FALSE} to say 
whether the sample size needed to achieve a certain power should be 
calculated. This calculation assumes a balanced design. By default, 
\code{sample_size_calc} is \code{FALSE}.}

\item{clusters_pr_group}{the number of clusters in each group. It should 
only be specified, if \code{sample_size_calc} is \code{TRUE}, and 
\code{cluster_size} is not specified.}

\item{cluster_size}{the number of measurements in each cluster. It should 
only be specified, if \code{sample_size_calc} is \code{TRUE}, and 
\code{clusters_pr_group} is not specified.}

\item{cluster_group_1}{a vector that contains the sizes of the clusters in 
group 1. If group 1 for instance has one cluster with 10 measurements, and 
three clusters with 12 measurements each, then we have to set 
\code{clusters_group_1} to \code{c(10,12,12,12)}.}

\item{cluster_group_2}{a vector that contains the sizes of the clusters in 
group 2.}
}
\value{
A list - If \code{sample_size_calc} is set to \code{FALSE}, then one the 
parameters \code{delta, sd, rho, power} must be left unspecified. The 
unspecified parameter will then be calculated by the function. If 
\code{sample_size_calc} is set to \code{TRUE}, then all of the parameters 
\code{delta, sd, rho, power} must be specified, but \code{clusters_group_1} 
and \code{clusters_group_2} must not be specified. On top of that, either 
\code{clusters_pr_group} or \code{cluster_size} must be specified. The 
other one will then be calculated by the function. Whether or not 
\code{sample_size_calc} is \code{TRUE} or \code{FALSE}, the function returns 
a list of all of the arguments to the function and the calculated parameter 
(the argument left unspecified when calling the function)
}
\description{
Calculates either the power of the cluster adjusted t test or the value 
of some other parameter needed to obtain a specified power of the test. 
The calculated power is only exact, if the design is balanced (same numbers
of clusters in each group, and same number of responses in each cluster). 
See XXXX for a description of the assumptions and construction of the 
cluster adjusted t test and how to calculate its power.
}
\examples{
#*************************************************************************
# power calculation is balanced design:
power.t.test.clustered(clusters_group_1=c(20,20,20), 
clusters_group_2=c(20,20,20), delta=2,sd=1.5,alternative="one.sided")

# power calculation in unbalanced design:
power.t.test.clustered(clusters_group_1=c(22,20,24), 
clusters_group_2=c(19,21,20,23), delta=2,sd=1.5,alternative="one.sided")

# calculation of the minimum treatment effect needed to 
# obtain a power of 0.8:
power.t.test.clustered(clusters_group_1=c(22,20,24), 
clusters_group_2=c(19,21,20,23), sd=1.5, power=0.8, alternative="one.sided")

# calculation of the needed number of measurements in each cluster
# to obtain a power of 0.8:
power.t.test.clustered(delta=2, sd=1.5, power=0.8, alternative="one.sided", 
sample_size_calc=TRUE, clusters_pr_group=5)

# calculation of the needed number of clusters in each group 
# to obtain a power of 0.8:
power.t.test.clustered(delta=2, sd=1.5, alternative="one.sided", 
sample_size_calc=TRUE, cluster_size=10)
#*************************************************************************
}
\seealso{
\code{\link{t.test.clustered.pval}} for p value of the cluster 
adjusted t test, \code{\link{t.test.clustered.stat}} for value of the 
cluster adjusted t statistic and \code{\link{simulate.power.t.test.clustered}}
for power simulation of the cluster adjusted t test.
}

