% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_t_test_clustered.R
\name{simulate.power.t.test.clustered}
\alias{simulate.power.t.test.clustered}
\title{Simulated power of the t test for cluster randomized designs}
\usage{
\method{simulate}{power.t.test.clustered}(num_experiments = 1000,
  clusters_group_1, clusters_group_2, delta, sd, rho, alternative,
  sig_level = 0.05)
}
\arguments{
\item{num_experiments}{is the number of experiments in the simulation.}

\item{delta}{the difference between the group means (the treatment effect).}

\item{sd}{the overall standard deviation of the responses, which is assumed 
to be equal in both groups.}

\item{rho}{the intracluster correlation coeffiecient, which is assumed to 
be equal in both groups.}

\item{alternative}{has to be set to one of two strings, either 
\code{"one.sided"} or \code{"two.sided"}, to say whether the alternative 
hypothesis is one-sided or two-sided.}

\item{sig_level}{the significance level of the t test. If it is not 
specified, it has default value \code{0.05}.}
}
\value{
A number - simulated power of the t test for cluster randomized
designs
}
\description{
The function simulates "\code{num_experiments} instances of data according 
to the model explained in XXXX with parameters defined by the arguments of 
the function. For each one of the instances, the function uses the t test 
for cluster randomized designs to test the null hypothesis that the group 
means are equal. The function outputs the simulated power of the test, 
which is the proportion of test rejecting the null hypothesis out of the 
total number of test.
}
\examples{
#*************************************************************************
# power simulation
simulate.power.t.test.clustered(num_experiments=1000, clusters_group_1 = c(20,25,15,20), 
                 clusters_group_2=c(15,18,25), delta=2, sd=1.5, rho=0.4, 
                 alternative="one.sided", sig_level=0.05) 
#*************************************************************************
}
\seealso{
\code{\link{t.test.clustered.pval}} for p value of the cluster 
adjusted t test, \code{\link{t.test.clustered.stat}} for value of the 
cluster adjusted t statistic and \code{\link{power.t.test.clustered}}
for power calculation of the cluster adjusted t test.
}

